updated features section
