--- src/assets/data/pricing/index.js
@@ -1,83 +1,0 @@
-
-const data = {
-    pricings: [
-        {
-            name: "Lifecycle",
-            description: "Manage any service mesh",
-            link: "#",
-            btnText: "Get Started",
-            services:[
-                {
-                    content: "Consul, Istio, Linkerd, Network Service Mesh, Open Service Mesh, App Mesh, Kuma, Maesh, Citrix Service Mesh, Octarine",
-                    className: "--on"
-                },
-                {
-                    content: "Explore service mesh functionality with 10 different sample applications",
-                    className: "--on"
-                },
-                {
-                    content: "Validate each service mesh's compliance with Service Mesh Interface (SMI) specifications",
-                    className: "--on"
-                }
-            ],
-            count: {
-                value: "11",
-                description: "service meshes supported"
-            }
-        },{
-            name: "Configuration",
-            description: "Operators with best practices",
-            link: "#",
-            btnText: "Get Started",
-            services:[
-                {
-                    content: "Built-in industry patterns and known best practices",
-                    className: "--on"
-                },
-                {
-                    content: "Apply custom configuration to your service mesh",
-                    className: "--on"
-                },
-                {
-                    content: "Advanced reporting",
-                    className: "--on"
-                }
-            ],
-            count: {
-                value: "1,000",
-                description: "tracked users"
-            }
-        },
-        {
-            name: "Performance",
-            description: "Assess the performance of your service mesh",
-            link: "#",
-            btnText: "Get Started",
-            services:[
-                {
-                    content: "Distributed performance analysis supporting 3 different types of load generators",
-                    className: "--on"
-                },
-                {
-                    content: "Built-in statistical analysis of service latency quartiles",
-                    className: "--on"
-                },
-                {
-                    content: "Integration with Prometheus and Grafana",
-                    className: "--on"
-                },
-                {
-                    content: "Baseline and track deltas of workload performance over time.",
-                    className: "--on"
-                }
-            ],
-            count: {
-                value: "1,113",
-                description: "performance tests run"
-            }
-        }
-      
-    ]
-    
-};
-export default data;
